requirements.txt
APScheduler==3.10.4
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
exceptiongroup==1.2.2
Flask==2.3.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
PySocks==1.7.1
pytz==2024.2
PyYAML==6.0.2
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
